[TPU VM] Avoid infinite recursion when cleaning up spot TPU VM (#1555)
* Terminate TPU VM in refresh without using backend

* Fix log path

* address comments

* rename argument

* Update sky/backends/backend_utils.py
